Added a flag for logging X-Forwarded... headers
Added a flag to enable X-Forwarding headers when the server is running behind a proxy. It is important that this remains off by default, since if the server is not behind a proxy which sets these headers, then a hacker could be set to '\7' for example, causing the terminal to beep whenever a log file is read. Additionally, the addition of the flag argument is done via variadic arguments, so that the API is backwards-compatible.